update: Repository ID Integer로 수정에 따른 코드 변경

+++ src/main/java/com/twoPotatoes/bobJoying/member/repository/TeamRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.twoPotatoes.bobJoying.member.entity.Team;
 
 @Repository
-public interface TeamRepository extends JpaRepository<Team, Long> {
+public interface TeamRepository extends JpaRepository<Team, Integer> {
 }
